Inheritable is only useful in pool plugin
